DO NOT MERGE: Fix usb host 2 transceiver init.
IPS1504 has a broken/non-usable reset. We need to have a special
initialisation procedure where we setup the pull-up/down then
enable it in order to boot it's PLL, then disable it until
the ULPI state machine on the imx31 is enabled.

With this voodoo procedure the transceiver seems to get initialised
correctly.

This is needed only on usb host 2 because the IMX31 do not enable
usb pinmux configuration by default on this host.